u2mfn: add API to use specific u2mfn FD
This enable to get multiple pages (via u2mfn_alloc_kpage) - kernel module holds
pointer to it in per-FD struct, so the same FD can't be reused for the next request.

More elegant solution should involve kernel module modification (stored page
list or sth), but it is planed to replace this part of code with libxenvchan
(or even more generic version), so this temporary solution should be enough.