notify-updates: ignore chown error
If running as normal user, chown will fail. This isn't a problem,
because the file is probably already owned by the correct user. The
whole point about this chown is to give access to the file for normal
user, so if the write succeeded, it isn't needed.

(cherry picked from commit 063b436b037f0e7e1725e5bdcca9b2bb1e692fac)