Added qubes-label entry to qubesdb on VM start
There have been a handful of requests for access to the Qube label
color from within the Qube itself for theming purposes. Adding this
entry to QubesDB will facilitate that. I do not believe this causes
a breach of security, as I feel the label should be treated the same
as the Qube name, which is also set in 000QubesVm.py.