Fix HAP (and PVH) detection
Libvirt lists the HAP feature even if it's not supported by the
hardware. So instead we get it from xen_caps (exposed there by a custom
patch). While we are at it also check for PVH support instead of HVM
support (the Xen code seems to requires a few more vmx features for PVH
than for HVM).