tests: detect never awaited coroutine objects
Python 3.7+ generates a RuntimeWarning when a native (i.e. defined with
'async def') coroutine object is never awaited, which usually points to
a bug.

Unfortunately, the RuntimeWarning happens during object finalization,
where it's impossible to raise it as an exception right away with
'warning.filterwarnings("error", ...)': That would just result in
'Exception ignored' messages. The workaround here is to accumulate all
the relevant warnings from some chunk of code and then raise an
exception afterwards.

Garbage collection can be triggered before and after that chunk of code,
for precise results if a coro is finalized as part of an object cycle.
But not by default, because with lots of quick tests it's expensive* to
gc in every part of the test (setUp(), the test method, tearDown(), and
every cleanup method). At least QubesTestCase.cleanup_gc(), which is
called by the decorated _callCleanup(), will already trigger gc at the
very end of each test.

* Results of './run-tests; time ./run-tests' in my dev VM:

  ~ .1% longer real time by default

  ~ 42% longer real time with gc_after
  (to prevent false _negatives_ - warnings that escape from a test part
  into the surrounding 'unittest' module code)

  ~ 85% longer real time with gc_after and gc_before
  (to also prevent false _positives_ - warnings that escape from
  elsewhere into a test part)